Removed #resForm margin formatting
Labels and data inputs are not lining up as intended, so I removed the margin formatting until I find a solution. This will keep everything in line, albeit the labels are far from the inputs. Also, I found an extra space between numbers and their units in a few areas; the spaces were removed.